scripts/requirements-screen-monitor.txt
mss==10.1.0
Pillow==11.2.1
pytesseract==0.3.13
openai>=1.40.0
